解决带父子关系时查看详情父对象long未转换为字符串的bug
